+++ esp-lp-hal/src/gpio.rs
@@ -1,34 +1,58 @@
-//! Low-power GPIO driver
+//! # General Purpose Input/Output
+//!
+//! ## Overview
 //!
 //! It's assumed that GPIOs are already configured correctly by the HP core.
+//!
+//! This driver supports various operations on GPIO pins, primarily manipulating
+//! the pin state (setting high/low, toggling).
+//!
+//! This module also implements a number of traits from `embedded-hal` to
+//! provide a common interface for GPIO pins.
+//!
+//! ## Examples
+//!
+//! ```rust
+//! fn main(gpio0: Input<0>, gpio1: Output<1>) -> ! {
+//!     loop {
+//!         let input_state: bool = gpio0.input_state();
+//!         gpio.set_output(input_state);
+//!
+//!         esp_lp_hal::delay::Delay.delay_millis(50);
+//!     }
+//! }
+//! ```
 
-#[cfg(feature = "esp32c6")]
-type LpIo = crate::pac::LP_IO;
-#[cfg(any(feature = "esp32s2", feature = "esp32s3"))]
-type LpIo = crate::pac::RTC_IO;
+cfg_if::cfg_if! {
+    if #[cfg(feature = "esp32c6")] {
+        type LpIo = crate::pac::LP_IO;
+        const MAX_GPIO_PIN: u8 = 7;
+    } else {
+        type LpIo = crate::pac::RTC_IO;
+        const MAX_GPIO_PIN: u8 = 21;
+    }
+}
 
-#[cfg(feature = "esp32c6")]
-const MAX_GPIO_PIN: u8 = 7;
-#[cfg(any(feature = "esp32s2", feature = "esp32s3"))]
-const MAX_GPIO_PIN: u8 = 21;
-
-#[non_exhaustive]
-pub struct Input<const PIN: u8> {}
+/// GPIO input driver
+pub struct Input<const PIN: u8>;
 
 impl<const PIN: u8> Input<PIN> {
+    /// Read the input state/level of the pin.
     pub fn input_state(&self) -> bool {
         unsafe { &*LpIo::PTR }.in_().read().bits() >> PIN & 0x1 != 0
     }
 }
 
-#[non_exhaustive]
-pub struct Output<const PIN: u8> {}
+/// GPIO output driver
+pub struct Output<const PIN: u8>;
 
 impl<const PIN: u8> Output<PIN> {
+    /// Read the output state/level of the pin.
     pub fn output_state(&self) -> bool {
         unsafe { &*LpIo::PTR }.out().read().bits() >> PIN & 0x1 != 0
     }
 
+    /// Set the output state/level of the pin.
     pub fn set_output(&mut self, on: bool) {
         if on {
             unsafe { &*LpIo::PTR }
@@ -48,7 +72,7 @@ pub unsafe fn conjure_output<const PIN: u8>() -> Option<Output<PIN>> {
     if PIN > MAX_GPIO_PIN {
         None
     } else {
-        Some(Output {})
+        Some(Output)
     }
 }
 
@@ -58,7 +82,7 @@ pub unsafe fn conjure_input<const PIN: u8>() -> Option<Input<PIN>> {
     if PIN > MAX_GPIO_PIN {
         None
     } else {
-        Some(Input {})
+        Some(Input)
     }
 }
 
@@ -104,18 +128,18 @@ impl<const PIN: u8> embedded_hal_02::digital::v2::StatefulOutputPin for Output<P
 #[cfg(feature = "embedded-hal-02")]
 impl<const PIN: u8> embedded_hal_02::digital::v2::toggleable::Default for Output<PIN> {}
 
-#[cfg(feature = "embedded-hal-1")]
-impl<const PIN: u8> embedded_hal_1::digital::ErrorType for Input<PIN> {
+#[cfg(feature = "embedded-hal")]
+impl<const PIN: u8> embedded_hal::digital::ErrorType for Input<PIN> {
     type Error = core::convert::Infallible;
 }
 
-#[cfg(feature = "embedded-hal-1")]
-impl<const PIN: u8> embedded_hal_1::digital::ErrorType for Output<PIN> {
+#[cfg(feature = "embedded-hal")]
+impl<const PIN: u8> embedded_hal::digital::ErrorType for Output<PIN> {
     type Error = core::convert::Infallible;
 }
 
-#[cfg(feature = "embedded-hal-1")]
-impl<const PIN: u8> embedded_hal_1::digital::InputPin for Input<PIN> {
+#[cfg(feature = "embedded-hal")]
+impl<const PIN: u8> embedded_hal::digital::InputPin for Input<PIN> {
     fn is_high(&mut self) -> Result<bool, Self::Error> {
         Ok(self.input_state())
     }
@@ -125,8 +149,8 @@ impl<const PIN: u8> embedded_hal_1::digital::InputPin for Input<PIN> {
     }
 }
 
-#[cfg(feature = "embedded-hal-1")]
-impl<const PIN: u8> embedded_hal_1::digital::OutputPin for Output<PIN> {
+#[cfg(feature = "embedded-hal")]
+impl<const PIN: u8> embedded_hal::digital::OutputPin for Output<PIN> {
     fn set_low(&mut self) -> Result<(), Self::Error> {
         self.set_output(false);
         Ok(())
@@ -138,8 +162,8 @@ impl<const PIN: u8> embedded_hal_1::digital::OutputPin for Output<PIN> {
     }
 }
 
-#[cfg(feature = "embedded-hal-1")]
-impl<const PIN: u8> embedded_hal_1::digital::StatefulOutputPin for Output<PIN> {
+#[cfg(feature = "embedded-hal")]
+impl<const PIN: u8> embedded_hal::digital::StatefulOutputPin for Output<PIN> {
     fn is_set_high(&mut self) -> Result<bool, Self::Error> {
         Ok(self.output_state())
     }